# server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AdminFlow API",
    "version": "2.4.0",
    "description": "Documentación oficial AdminFlow. El sistema utiliza MongoDB como motor de base de datos exclusivo."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Back-end API"
    }
  ],
...
  "paths": {
...
    "/contracts": {
      "get": {
        "tags": [
          "Contracts"
        ],
        "summary": "Listar contratos",
        "responses": {
          "200": {
            "description": "Lista de contratos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Contract"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Contracts"
        ],
        "summary": "Crear nuevo contrato",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Contract"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contrato creado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contract"
                }
              }
            }
          }
        }
      }
    },
    "/contracts/{id}": {
      "get": {
        "tags": [
          "Contracts"
        ],
        "summary": "Obtener contrato",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Contrato",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contract"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Contracts"
        ],
        "summary": "Actualizar contrato",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Contract"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contrato actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contract"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Contracts"
        ],
        "summary": "Eliminar contrato",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Contrato eliminado"
          }
        }
      }
    },
    "/api/push/subscribe": {
      "post": {
        "tags": [
          "Push Notifications"
        ],
        "summary": "Suscribir a notificaciones push",
        "description": "Registra una suscripción push para recibir notificaciones de escritorio",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "endpoint"
                ],
                "properties": {
                  "endpoint": {
                    "type": "string",
                    "description": "URL del endpoint push del navegador"
                  },
                  "keys": {
                    "type": "object",
                    "properties": {
                      "p256dh": {
                        "type": "string"
                      },
                      "auth": {
                        "type": "string"
                      }
                    }
                  },
                  "expirationTime": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Suscripción registrada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos de suscripción inválidos"
          }
        }
      }
    },
    "/api/push/unsubscribe": {
      "post": {
        "tags": [
          "Push Notifications"
        ],
        "summary": "Desuscribir de notificaciones push",
        "description": "Elimina una suscripción push existente",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "endpoint"
                ],
                "properties": {
                  "endpoint": {
                    "type": "string",
                    "description": "URL del endpoint push a eliminar"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Suscripción eliminada exitosamente"
          }
        }
      }
    },
...
    "/api/push/send": {
      "post": {
        "tags": [
          "Push Notifications"
        ],
        "summary": "Enviar notificación push",
        "description": "Envía una notificación push real a todos los dispositivos suscritos usando web-push",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "Título de la notificación"
                  },
                  "body": {
                    "type": "string",
                    "description": "Cuerpo del mensaje"
                  },
                  "url": {
                    "type": "string",
                    "description": "URL a abrir al hacer clic"
                  },
                  "tag": {
                    "type": "string",
                    "description": "Tag para agrupar notificaciones"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Notificación enviada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "payload": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Contract": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
    }
  }
}